Move image types to image mod
